requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
pandas
numpy
scipy
scanpy
pydeseq2
anndata
requests
openpyxl
gunicorn
python-dotenv
